requirements.txt
rdkit-pypi==2023.3.1b1
scikit-learn>=1.0.2
pandas>=1.3.0
matplotlib>=3.4.3
streamlit>=1.28.0
py3Dmol>=2.0.0
